+++ packages/components/CHANGELOG.md
@@ -14,6 +14,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 
 ## [Unreleased]
 
+### Changed
+
+- `Flyout`, `Modal`, `Sheet`: Extracted shared `<dialog>` render structure into internal `DialogBase` functional
+  component to reduce duplication across all three components ([#4364](https://github.com/porsche-design-system/porsche-design-system/pull/4364))
+
 ## [4.1.0] - 2026-05-06
 
 ## [4.1.0-rc.0] - 2026-05-05